update
1. Create AccountHandler to deal with account settings
2. Update thread_subscribe()
