+++ trader/utils/__init__.py
@@ -60,14 +60,17 @@ def create_queue(target_list, crawled_list=[]):
     return (q)
 
 
-def get_contract(target: str):
-    if target[:3] in API.Contracts.Indexs.__dict__:
-        return API.Contracts.Indexs[target[:3]][target]
-    elif target[:3] in API.Contracts.Futures.__dict__:
-        return API.Contracts.Futures[target[:3]][target]
-    elif target[:3] in API.Contracts.Options.__dict__:
-        return API.Contracts.Options[target[:3]][target]
-    return API.Contracts.Stocks[target]
+def get_contract(target: str, api=None):
+    if api is None:
+        api = API
+
+    if target[:3] in api.Contracts.Indexs.__dict__:
+        return api.Contracts.Indexs[target[:3]][target]
+    elif target[:3] in api.Contracts.Futures.__dict__:
+        return api.Contracts.Futures[target[:3]][target]
+    elif target[:3] in api.Contracts.Options.__dict__:
+        return api.Contracts.Options[target[:3]][target]
+    return api.Contracts.Stocks[target]
 
 
 def concat_df(df1: pd.DataFrame, df2: pd.DataFrame, sort_by=[], reset_index=False):
