+++ trader/utils/objects/__init__.py
@@ -1,0 +1,3 @@
+class Margin:
+    available_margin = 0
+    equity = 0
